# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: Democra — API interna (server/)
  version: "1.0.0"
  description: >
    Documentación base de la API Express de Democra (server/index.js).
    Extraída de los routers reales (server/routes/*.js), no inventada.
    Todos los endpoints requieren un Bearer token de Supabase Auth salvo
    donde se indique explícitamente lo contrario (terminal-login, validate-ruc).
    Este archivo es el punto de partida ("estructura base") para mantener la
    documentación de la API sincronizada con el código — se debe actualizar
    cada vez que se agregue, cambie o elimine un endpoint en server/routes/.
servers:
  - url: http://localhost:8787/api
    description: Desarrollo local (server/index.js, puerto por defecto de config.js)
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: Supabase JWT
      description: >
        Access token de una sesión de Supabase Auth. El backend lo resuelve
        vía resolveAuthContext() (server/supabase.js) para obtener user + profile.
  schemas:
    ErrorResponse:
      type: object
      description: Formato estándar de error (server/utils/http.js sendError()).
      properties:
        error_code:
          type: string
          example: "IAM-004"
        error_type:
          type: string
          example: "auth"
        message:
          type: string
...
    Sede:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        is_active:
          type: boolean
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
  responses:
    Unauthorized:
      description: Token ausente, inválido o expirado.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    ValidationError:
      description: Body o parámetros inválidos.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
security:
  - bearerAuth: []
paths:
...
  /sedes:
    get:
      tags: [sedes]
      summary: Lista las sedes del tenant actual
      responses:
        "200":
          description: Lista de sedes.
          content:
            application/json:
              schema:
                type: object
                properties:
                  sedes:
                    type: array
                    items:
                      $ref: "#/components/schemas/Sede"
        "401":
          $ref: "#/components/responses/Unauthorized"
    post:
      tags: [sedes]
      summary: Crea una nueva sede (requiere ser administrador del tenant)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name:
                  type: string
      responses:
        "201":
          description: Sede creada.
          content:
            application/json:
              schema:
                type: object
                properties:
                  sede:
                    $ref: "#/components/schemas/Sede"
        "400":
          $ref: "#/components/responses/ValidationError"
        "403":
          description: El actor no es administrador del tenant (IAM-003).
  /sedes/{sedeId}:
    put:
      tags: [sedes]
      summary: Actualiza el nombre o estado de una sede (requiere ser administrador)
      parameters:
        - name: sedeId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                is_active:
                  type: boolean
      responses:
        "200":
          description: Sede actualizada.
          content:
            application/json:
              schema:
                type: object
                properties:
                  sede:
                    $ref: "#/components/schemas/Sede"
        "403":
          description: El actor no es administrador del tenant (IAM-003).
        "404":
          description: Sede no encontrada en este tenant.
    delete:
      tags: [sedes]
      summary: Desactiva una sede (soft-delete — no elimina la fila; ver comentario en sedes.js)
      parameters:
        - name: sedeId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "204":
          description: Sede desactivada.
        "403":
          description: El actor no es administrador del tenant (IAM-003).
        "404":
          description: Sede no encontrada en este tenant.